added standalone .svg
Standalone SVG example. Not used anywhere in the project because you
can't reference elements within an svg file. It's better to build the
SVG in your html as an html element than to reference the svg file.
Gives you more power over its behavior.
Upside of a standalone SVG file is you can embed the SVG file as an
image through BB Codes.